fix: parser namespace semantics, decimal quantities, version constant, artifact hygiene
- Fix parse_role to skip in_keyword and capture domain identifier correctly (DF-008)
- Change Flow AST quantity from i32 to Decimal for precision preservation (DF-010)
- Make duplicate detection namespace-aware using (namespace, name) keys (DF-009)
- Fix VERSION constant to use env!("CARGO_PKG_VERSION") instead of hardcoded "0.1.0" (DF-021)
- Add resolve_by_name helper for unqualified reference resolution
- Flows now use Flow::new_with_namespace with correct namespace propagation
- Define generated artifact policy; update .gitignore for build outputs
- Remove tracked build outputs (pkg/*, fix_*.py, sea.tar.gz)
- Update ast_schema and ast_convert for Decimal quantity serialization

+++ sea-core/src/parser/printer.rs
@@ -2,6 +2,7 @@ use crate::parser::ast::{
     Ast, AstNode, FileMetadata, ImportDecl, ImportItem, ImportSpecifier, MappingRule,
     ProjectionOverride,
 };
+use rust_decimal::Decimal;
 use serde_json::Value as JsonValue;
 use std::collections::HashMap;
 use std::fmt::Write;
@@ -327,7 +328,7 @@ impl PrettyPrinter {
         annotations: &HashMap<String, JsonValue>,
         from: &str,
         to: &str,
-        quantity: Option<i32>,
+        quantity: Option<Decimal>,
     ) -> String {
         let mut lines = Vec::new();
         let head = format!("Flow {}", self.quote(resource));
